pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dreamcatcher-memory"
version = "0.4.0"
description = "Living memory for AI agents — nightly re-fine-tuning of compact models on personal data"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Alex Skatell"}]
keywords = ["ai", "agents", "memory", "fine-tuning", "personal-ai", "dreamcatcher", "living-memory"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Minimal dependencies for the server + client (no GPU required)
dependencies = [
    "pyyaml>=6.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
# PyTorch training (NVIDIA GPU / CPU)
train = [
    "torch>=2.2.0",
    "transformers>=4.51.0",
    "datasets>=2.19.0",
    "accelerate>=0.28.0",
    "safetensors>=0.4.0",
    "anthropic>=0.25.0",
]
# MLX training (Apple Silicon) — install separately since mlx is platform-specific
# pip install mlx mlx-lm
# For OpenAI as extraction provider instead of Anthropic
openai = ["openai>=1.20.0"]
# Claude Code MCP integration
claude-code = ["mcp>=1.0.0"]
# Everything (PyTorch path)
all = ["dreamcatcher-memory[train,openai,claude-code]"]

[project.scripts]
dreamcatcher = "dreamcatcher.__main__:main"

[project.urls]
Homepage = "https://github.com/alexskatell/living-memory"
Documentation = "https://github.com/alexskatell/living-memory#readme"
Repository = "https://github.com/alexskatell/living-memory"

[tool.setuptools.packages.find]
include = ["dreamcatcher*"]
